+++ src/microservices/Microservices.IsIdentifiable/Runners/DicomFileRunner.cs
@@ -120,7 +120,10 @@ namespace Microservices.IsIdentifiable.Runners
                     ValidateDicomItem(fi, dicomFile, dataSet, dicomItem);
             }
             else
+            {
                 _logger.Info("File does not contain valid preamble and header: " + fi.FullName);
+                throw new ApplicationException("File does not contain valid preamble and header: " + fi.FullName);
+            }
 
             DoneRows(1);
         }
@@ -142,7 +145,7 @@ namespace Microservices.IsIdentifiable.Runners
                 {
                     value = DicomTypeTranslaterReader.GetCSharpValue(dataset, dicomItem);
                 }
-                catch (System.FormatException e)
+                catch (System.FormatException)
                 {
                     value = "Unknown value for "+dicomItem;
                 }
@@ -244,7 +247,7 @@ namespace Microservices.IsIdentifiable.Runners
             {
                 // An internal error should cause IsIdentifiable to exit
                 _logger.Info(e, "Could not run Tesseract on '" + fi.FullName + "'");
-                throw new Exception ("Could not run Tesseract on '" + fi.FullName + "'", e);
+                throw new ApplicationException ("Could not run Tesseract on '" + fi.FullName + "'", e);
 
                 // OR add a message to the report saying we failed to run OCR
                 //string problemField = "PixelData";